WIP, fix: expiration validation (#4)
* chore: add missing test hook for calculate-expiration-offset

* chore(polyfill): clean non-digits in parse-int

* refactor(core): convert maxnumber to int

* fix(encoding): handle single-parameter salts (split on ?)

* refactor: decouple challenge expiration logic from current system clock

* fix: accidental mutation of `:expires`

When the challenge was first created, `:expires` referenced the offset from current/base timestamp in seconds.
But it was accidentally shadowed inside `create-challenge` and upon being re-submitted for validation, it
referenced the result of adding this offset to the base timestamp, leading to a repeated calculation with 2 large values
when recreating a control challenge. The solution was to introduce a TTL parameter, allowing to easily make `expires`
and TTL itself immutable.